backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - PostgreSQL
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Vector Database - Qdrant
qdrant-client==1.7.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
cryptography==42.0.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Clients
httpx==0.26.0
requests==2.31.0

# Machine Learning - LSTM & Embeddings
torch==2.1.2
numpy==1.26.3
tensorflow==2.15.0
scikit-learn==1.4.0
sentence-transformers==2.3.1

# NLP & Sentiment
vaderSentiment==3.3.2

# Utilities
python-dateutil==2.8.2
pytz==2024.1
loguru==0.7.2

# Communication
twilio==8.11.1

# Server
gunicorn==21.2.0

